Add files via upload
The primary objective of this project is to develop a robust and clinically applicable stroke prediction model using patient data.

Once validated and tested, this predictive model will become integral to the organization's clinical decision-making process. It will enable the identification of patients at high risk of stroke, facilitating early intervention and prevention measures. Moreover, the model will track the progress of high-risk patients and monitor the effectiveness of preventive measures in reducing stroke incidence.